Added Job.list method
Added the entity Job and its only method: list. My plan is to return Job
entities in the future instead of a hash but for now is similar to Genre.list

I have also added it to the README file and updated the links to the sections
since the way it was you needed to reload the browser to reach the
sections.

I have also added a missing spec file in themoviedb.gemspec