Change People entity to Person
In the documentation of TMDB we can see People in the list of sections
and after that every request is based on person, like person/3 or
person/popular.

Because of this, a change in this gem to reflect that could be positive.
Besides that, the methods 'popular' and 'latest' were not working before this
change because they were using the paths people/popular and
people/latest, this commit fixes that too.

Specs for this entity have been added too.