Allow searching movies with filters
The [API for Movies](https://developers.themoviedb.org/3/search/search-movies)
allows searching with more than just the query. However, the gem didn't
have a way of adding this filters.

This commit allows passing a hash of filters to the search/find method
so these parameters can be added when searching.